feat(home): implement Senior Dev corrections
- Converted Button to a reusable Link component to properly use the <a> tag where needed.
- Renamed: public/heroimg.jpg → public/images/heroimg.jpg to support blur image implementation.
- Updated layout from flex to grid for improved responsive width handling.
